backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pypdf==6.1.2
Werkzeug==3.1.3
flask-cors
gunicorn

python-magic>=0.4.27
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
Pillow>=10.0.0